requirements.txt
fastapi[standard]
SQLAlchemy
uvicorn[standard]
alembic
passlib[bcrypt]
openai
python-dotenv
pydantic
pydantic-settings
logging
groq